+++ publish.sh
@@ -1,0 +1,9 @@
+version=`node -e 'console.log(require("./package.json").version)'`
+echo "version $version"
+
+if [ "$version" != "" ]; then
+    git tag -a "$version" -m "`git log -1 --format=%s`"
+    echo "Created a new tag, $version"
+    git push origin --tags
+    npm publish --access public
+fi